+++ train.py
@@ -1,13 +1,12 @@
-"""
-Entry point to tune, train and evaluate models. Tuning uses K-Fold
-cross validation to estimate model performance.
-"""
+"""Entry point to model training and evaluation."""
 from argparse import ArgumentParser, BooleanOptionalAction
 
 from superphot_plus.file_paths import INPUT_CSVS, PROBS_FILE
-from superphot_plus.trainer import CrossValidationTrainer
+from superphot_plus.trainer import SuperphotTrainer
 
-if __name__ == "__main__":
+
+def extract_cmd_args():
+    """Extracts the script command-line arguments."""
     parser = ArgumentParser(
         description="Entry point to train and evaluate models using K-Fold cross validation",
     )
@@ -15,11 +14,6 @@ if __name__ == "__main__":
         "--input_csvs",
         help="List of CSVs containing light curve data (comma separated)",
         default=",".join(INPUT_CSVS),
-    )
-    parser.add_argument(
-        "--model_name",
-        help="If set, the trainer loads this model and skips tuning",
-        default=None,
     )
     parser.add_argument(
         "--sampler",
@@ -30,13 +24,8 @@ if __name__ == "__main__":
     parser.add_argument(
         "--include_redshift",
         help="If flag is set, include redshift data for training",
-        default=False,
+        default=True,
         action=BooleanOptionalAction,
-    )
-    parser.add_argument(
-        "--num_hp_samples",
-        help="Name of parameter combinations to try",
-        default=10,
     )
     parser.add_argument(
         "--extract_wc",
@@ -50,29 +39,31 @@ if __name__ == "__main__":
         default=PROBS_FILE,
     )
     parser.add_argument(
-        "--num_cpu",
-        help="Number of CPUs to use in each parallel experiment",
-        default=2,
+        "--config_name",
+        help="The name of the file containing the model configuration",
+        required=True,
     )
     parser.add_argument(
-        "--num_gpu",
-        help="Number of GPUs to use in each parallel experiment",
-        default=0,
+        "--load_checkpoint",
+        help="If set, load pretrained model for the respective configuration",
+        default=False,
+        action=BooleanOptionalAction,
     )
+    return parser.parse_args()
 
-    args = parser.parse_args()
 
-    trainer = CrossValidationTrainer(
-        model_name=args.model_name,
+if __name__ == "__main__":
+    args = extract_cmd_args()
+
+    trainer = SuperphotTrainer(
         sampler=args.sampler,
         include_redshift=args.include_redshift,
+        config_name=args.config_name,
         probs_file=args.probs_file,
     )
 
     trainer.run(
         input_csvs=args.input_csvs.split(","),
-        num_hp_samples=args.num_hp_samples,
         extract_wc=args.extract_wc,
-        num_cpu=args.num_cpu,
-        num_gpu=args.num_gpu,
+        load_checkpoint=args.load_checkpoint,
     )
